refactor(settings): extract SyncSettingUseCase to interface and implementation
